Move tests to test for the class containing the logic
